+++ R/mod_time_provinces.R
@@ -18,7 +18,6 @@ mod_time_provinces_ui <- function(id){
 
   f7Card(
     title = "Confirmed Cases - China",
-    f7Toggle(ns("log"), "Logarithmic Scale log(1 + x)", checked = TRUE),
     echarts4r::echarts4rOutput(ns("map"), height = "70vh"),
     footer = uiOutput(ns("copy_ui"))
   )
@@ -36,26 +35,34 @@ mod_time_provinces_server <- function(input, output, session, df){
   embed_url <- golem::get_golem_options("embed_url")
 
   output$copy_ui <- renderUI({
-    copy(embed_url, "jhu", paste0("&chart=timeline-provinces&log=", tolower(input$log)))
+    copy(embed_url, "jhu", "&chart=timeline-provinces")
   })
 
   output$map <- echarts4r::renderEcharts4r({
-    mod_time_provinces_echarts(df, input$log)
+    mod_time_provinces_echarts(df)
   })
 }
 
-mod_time_provinces_echarts <- function(df, log){
+mod_time_provinces_echarts <- function(df, log = FALSE){
   dat <- df %>% 
     dplyr::filter(country %in% c("Mainland China", "Hong Kong", "Taiwan")) %>% 
     dplyr::left_join(chinese_provinces, by = "state") %>% 
     dplyr::arrange(desc(date)) %>% 
     dplyr::select(chinese, type, date, cases) %>% 
-    dplyr::mutate(
-      cases = dplyr::case_when(
-        log ~ log1p(cases),
-        TRUE ~ cases
+    dplyr::filter(type == "confirmed")
+
+  x <- dplyr::arrange(dat, cases) %>% dplyr::pull(cases) %>% unique()
+  n <- 5
+
+  pieces <- split(x, sort(x%%n)) %>% 
+    unname() %>% 
+    purrr::map(range) %>% 
+    purrr::map(function(x){
+      list(
+        gt = round_up(x[1]),
+        lte = round_up(x[2])
       )
-    )
+    })
 
   index <- length(unique(dat$date)) -1 
   titles <- unique(dat$date) %>% 
@@ -64,17 +71,17 @@ mod_time_provinces_echarts <- function(df, log){
       list(text = format(x, "%d %B %H:00"))
     })
 
-  mx <- max(dat$cases)
-  mn <- min(dat$cases)
-
   dat %>% 
     dplyr::group_by(date) %>% 
     echarts4r::e_charts(chinese, timeline = TRUE) %>% 
     echarts4r.maps::em_map("China") %>% 
     echarts4r::e_map(cases, map = "China", name = "confirmed") %>% 
     echarts4r::e_visual_map(
-      max = mx,
-      min = mn,
+      type = "piecewise",
+      pieces = pieces,
+      outOfRange = list(
+        color = "#000"
+      ),
       orient = "horizontal",
       right = "center",
       top = 25,